Publishing to gh-pages

+++ build/Build-Site.ps1
@@ -1,9 +1,14 @@
+param (
+    [switch]
+    $CleanUp,
+    [switch]
+    $Push
+)
 # copy documentation to output folder
 copy-item -Path "$PSScriptRoot/../documentation/*.md" -Destination "$PSScriptRoot/../pages/content/cmdlets"
 
 $items = Get-ChildItem -Path "$PSScriptRoot/../pages/content/cmdlets"
-foreach($item in $items)
-{
+foreach ($item in $items) {
     $metadata = Get-MarkdownMetadata -Path $item
     $title = $metadata.title
     $content = Get-Content -Path $item.FullName -Raw
@@ -11,15 +16,19 @@ foreach($item in $items)
     $header = "---
 title: $title
 ---"
-    $pageContent = $splitted[2] -replace "# $title",""  # Remove Title markdown
+    $pageContent = $splitted[2] -replace "# $title", ""  # Remove Title markdown
     $newContent = $header + $pageContent
     $newContent | Out-File $item.FullName -Force
 }
-hugo.exe -s "$PSScriptRoot\..\pages" -d "$PSScriptRoot\..\public"
+hugo.exe -s "$PSScriptRoot\..\pages" -d "$PSScriptRoot\..\public" --cleanDestinationDir
 
-# Remove cmdlets folder in pages folder. No need for duplicates here.
-Remove-Item "$PSScriptRoot/../pages/content/cmdlets/*" -Recurse -Force -ErrorAction Stop
-Set-Location $PSScriptRoot/../public 
-git add --all 
-git commit -m "Publishing to gh-pages"
-git push origin
+if ($CleanUp) {
+    # Remove cmdlets folder in pages folder. No need for duplicates here.
+    Remove-Item "$PSScriptRoot/../pages/content/cmdlets/*" -Recurse -Force -ErrorAction Stop
+}
+if ($Push) {
+    Set-Location $PSScriptRoot/../public 
+    git add --all 
+    git commit -m "Publishing to gh-pages"
+    git push origin
+}
